+++ rs/tests/consensus/Cargo.toml
@@ -11,19 +11,26 @@ anyhow = { workspace = true }
 candid = { workspace = true }
 canister-test = { path = "../../rust_canisters/canister_test" }
 futures = { workspace = true }
+ic-agent = { workspace = true }
 ic-base-types = { path = "../../types/base_types" }
 ic_consensus_system_test_utils = { path = "./utils" }
+ic-crypto-test-utils-reproducible-rng = { path = "../../crypto/test_utils/reproducible_rng" }
 ic-nns-constants = { path = "../../nns/constants" }
 ic-nns-governance = { path = "../../nns/governance" }
 ic-recovery = { path = "../../recovery" }
 ic-registry-subnet-type = { path = "../../registry/subnet_type" }
 ic-system-test-driver = { path = "../driver" }
 ic-types = { path = "../../types/types" }
+ic-universal-canister = { path = "../../universal_canister/lib" }
+k256 = { workspace = true }
 rand = { workspace = true }
 rand_chacha = { workspace = true }
 registry-canister = { path = "../../registry/canister" }
+reqwest = { workspace = true }
+serde_cbor = { workspace = true }
 slog = { workspace = true }
 tokio = { workspace = true }
+url = { workspace = true }
 
 [[bin]]
 name = "ic-systest-consensus-performance"
@@ -40,3 +47,20 @@ path = "node_graceful_leaving_test.rs"
 [[bin]]
 name = "ic-systest-replica-determinism-test"
 path = "replica_determinism_test.rs"
+
+[[bin]]
+name = "ic-systest-dual-workload-test"
+path = "dual_workload_test.rs"
+
+[[bin]]
+name = "ic-systest-max-ingress-payload-size-test"
+path = "max_ingress_payload_size_test.rs"
+
+[[bin]]
+name = "ic-systest-max-net-payload-size-test"
+path = "max_xnet_payload_size_test.rs"
+
+[[bin]]
+name = "ic-systest-request-auth-malicious-replica-test"
+path = "request_auth_malicious_replica_test.rs"
+
